requirements.txt
opencv-python
pymodbus
zbar
numpy